full update

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: mobouifr <mobouifr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/26 15:58:34 by mobouifr          #+#    #+#             */
-/*   Updated: 2024/08/21 16:45:07 by mobouifr         ###   ########.fr       */
+/*   Updated: 2024/08/29 17:47:02 by mobouifr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ int		ft_strncmp(const char *s1, const char *s2, size_t n);
 void	*ft_memchr(const void *s, int c, size_t n);
 int		ft_memcmp(const void *s1, const void *s2, size_t n);
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len);
-int		ft_atoi(char *str, char **string, long *all_nums);
+int		ft_atoi(char *str, char **string, int *all_nums);
 void	*ft_calloc(size_t count, size_t size);
 char	*ft_strdup(const char *s1);
 /* part 2 function */
